+++ ICD.Connect.Displays.Sharp/Devices/Consumer/SharpDisplay.cs
@@ -238,7 +238,7 @@ namespace ICD.Connect.Displays.Sharp.Devices.Consumer
 		/// </summary>
 		private void KeepAliveCallback()
 		{
-			if (ConnectionStateManager.IsConnected && SerialQueue.CommandCount == 0)
+			if (ConnectionStateManager.IsConnected && SerialQueue != null && SerialQueue.CommandCount == 0)
 				SendCommand(SharpDisplayCommands.POWER_QUERY);
 		}