+++ ICD.Connect.Displays.Sharp/Properties/AssemblyInfo.cs
@@ -5,4 +5,3 @@ using System.Reflection;
 [assembly: AssemblyProduct("ICD.Connect.Displays.Sharp")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
 [assembly: AssemblyVersion("12.0.0.0")]
-