+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - MockDisplayWithAudio no longer inheriting from serial display implementations
 
+## [11.1.1] - 2019-01-28
+### Changed
+ - Failing gracefully when no SerialQueue is assigned to displays
+ - Fix for NullRef in PanasonicClassicDisplay
+
 ## [11.1.0] - 2019-01-15
 ### Added
  - Added Barco VideoWall display driver